Refactoring profile, changing time formatting
