+++ app/src/main/java/com/metropolia/eatthefrog/constants/Constants.kt
@@ -1,4 +1,6 @@
 package com.metropolia.eatthefrog.constants
 
-const val DATE_FORMAT = "dd.MM.yyyy"
+const val SHARED_PREF_KEY = "PREFERENCES_KEY"
+const val USERNAME_KEY = "USERNAME_KEY"
+const val DATE_FORMAT = "d.M.yyyy"
 const val TIME_FORMAT = "HH:mm"
