+++ app/src/main/java/com/metropolia/eatthefrog/screens/AddTaskScreen.kt
@@ -143,7 +143,7 @@ fun AddTaskScreenC(viewModel: AddTaskScreenViewModel, navHost: NavHostController
         context,
         { _, hour: Int, minute: Int ->
             sTime.value = String.format("%02d:%02d", hour, minute)
-        }, mHour, mMinute, false
+        }, mHour, mMinute, true
     )
 
     Column(
